+++ src/app/(views)/dashboard/(catalogs)/catalog-table.tsx
@@ -1,8 +1,20 @@
+import { AlertTriangle } from "lucide-react";
 import Link from "next/link";
+import { PropsWithChildren, ReactNode } from "react";
 
 import { DeleteIcon, EditIcon, LinkIcon } from "~/components/custom/icons";
 import JustTip from "~/components/custom/just-the-tip";
 import { Button } from "~/components/shadcn/button";
+import {
+  Dialog,
+  DialogClose,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+  DialogTrigger,
+} from "~/components/shadcn/dialog";
 import {
   Table,
   TableBody,
@@ -67,17 +79,14 @@ function CatalogTable({ catalogs, onDelete, onEdit }: CatalogTableProps) {
                     </Button>
                   </JustTip>
 
-                  <JustTip label="Remove Catalog">
-                    <Button
-                      variant="outline"
-                      onClick={() => onDelete(catalog?.id)}
-                    >
+                  <DeleteModal handleDelete={() => onDelete(catalog?.id)}>
+                    <Button variant="outline">
                       <DeleteIcon
                         size={24}
-                        className="text-red-700 hover:text-red-500 cursor-pointer"
+                        className="text-primary/70 hover:text-primary cursor-pointer"
                       />
                     </Button>
-                  </JustTip>
+                  </DeleteModal>
                 </div>
               </TableCell>
             </TableRow>
@@ -89,3 +98,38 @@ function CatalogTable({ catalogs, onDelete, onEdit }: CatalogTableProps) {
 }
 
 export default CatalogTable;
+
+type DeleteModalProps = {
+  handleDelete: () => void;
+  children: ReactNode;
+};
+
+function DeleteModal({ children, handleDelete }: DeleteModalProps) {
+  return (
+    <Dialog>
+      <DialogTrigger asChild>{children}</DialogTrigger>
+      <DialogContent>
+        <DialogHeader>
+          <DialogTitle className="flex items-center gap-2">
+            <AlertTriangle className="h-5 w-5 text-primary" />
+            Confirm Deletion
+          </DialogTitle>
+          <DialogDescription className="text-primary/70">
+            This action is irreversible. The catalog will be permanently
+            deleted.
+          </DialogDescription>
+        </DialogHeader>
+        <DialogFooter className="sm:justify-start">
+          <Button type="button" onClick={handleDelete}>
+            Delete
+          </Button>
+          <DialogClose>
+            <Button type="button" variant="outline">
+              Cancel
+            </Button>
+          </DialogClose>
+        </DialogFooter>
+      </DialogContent>
+    </Dialog>
+  );
+}