+++ src/components/custom/footer.tsx
@@ -17,25 +17,13 @@ export default function Footer() {
             </a>
           </p>
         </div>
-        <div className="flex gap-2 items-center text-xs text-gray-500">
-          <a
-            className="cursor-pointer hover:text-gray-300"
-            href="https://github.com/realChakrawarti/yt-catalog"
-            target="_blank"
-          >
-            <span className="flex gap-2 items-center tracking-tight">
-              Open Source on GitHub
-              <GithubIcon size={18} />
-            </span>
-          </a>
-          <a
-            className="cursor-pointer hover:text-gray-300"
-            href="https://github.com/realChakrawarti/yt-catalog/blob/main/CHANGELOG.md"
-            target="_blank"
-          >
-            <p className="tracking-wider">v{packageInfo.version}</p>
-          </a>
-        </div>
+        <a
+          className="cursor-pointer text-xs text-gray-500 hover:text-gray-300"
+          href="https://github.com/realChakrawarti/yt-catalog/blob/main/CHANGELOG.md"
+          target="_blank"
+        >
+          <p className="tracking-wider">v{packageInfo.version}</p>
+        </a>
       </div>
     </footer>
   );