+++ src/components/custom/header.tsx
@@ -7,7 +7,7 @@ import AppIcon from "../../../public/icon.png";
 import { useAuth } from "../../app/auth/context-provider";
 import { Avatar, AvatarFallback, AvatarImage } from "../shadcn/avatar";
 import { Button } from "../shadcn/button";
-import AuthButton from "./auth-buttons";
+import { GitHubStargazer } from "./github-stargazers";
 import { LogoutIcon } from "./icons";
 import JustTip from "./just-the-tip";
 import ThemeToggle from "./theme-toggle";
@@ -17,17 +17,35 @@ const Header = () => {
 
   return (
     <header className="h-14 sticky z-50 top-0 w-full border-b border-border/40 self-start flex justify-between items-center bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60 dark:border-border">
-      <div className="flex-1 flex justify-between items-end px-2 container mx-auto">
-        <Link href={user ? "/dashboard" : "/"}>
-          <h1 className="flex gap-1">
-            <Image src={AppIcon} alt="YTCatalog" className="size-7" />
-            <p className="self-end text-lg tracking-wide">YTCatalog</p>
-          </h1>
-        </Link>
+      <div className="flex-1 flex justify-between items-center px-2 container mx-auto">
+        <div className="flex gap-2 items-center">
+          <Link href="/">
+            <h1 className="flex gap-1">
+              <Image src={AppIcon} alt="YTCatalog" className="size-7" />
+              {user ? null : (
+                <p className="self-end text-lg tracking-wide">YTCatalog</p>
+              )}
+            </h1>
+          </Link>
+          {user ? (
+            <Link
+              className="text-primary/70 hover:text-primary"
+              href="/dashboard"
+            >
+              Dashboard
+            </Link>
+          ) : null}
+        </div>
         <div className="flex gap-3 items-center">
+          <Link
+            className="text-primary/70 hover:text-primary"
+            href="/explore"
+          >
+            Explore
+          </Link>
           {user ? (
             <>
-              <Avatar className="size-8 rounded-lg">
+              <Avatar className="size-8 rounded-lg hidden md:block">
                 <AvatarImage
                   src={
                     user?.photoURL ||
@@ -45,7 +63,7 @@ const Header = () => {
               </JustTip>
             </>
           ) : (
-            <AuthButton />
+            <GitHubStargazer owner="realChakrawarti" repo="yt-catalog" />
           )}
           <ThemeToggle />
         </div>